Java Encryption project
Implementation of Front-end(redactor form) for inserting text (UTF-8)
and key for encryption and decryption. I pass the text information using
ajax to the server side where AES algorithm is used for encryption and
decryption of the text. The result is returned back to the browser.
Apache Tomkat  is used for running server.